PHP 8.1.3 (#2)
- [ ] Panic Testing EngineDefine
- [x] Fixed Readme
- [x] Fixed Testing strlen(); (probably)
- [x] Added script to install or build from source on debian 
- [x] Dropped support for PHP < 8
- [x] Changed names for easier use. 
- [x] Dropped _*.c
