+++ value.go
@@ -86,7 +86,7 @@ func NewValue(val interface{}) (*Value, error) {
 		for i := 0; i < v.Len(); i++ {
 			vs, err := NewValue(v.Index(i).Interface())
 			if err != nil {
-				C._value_destroy(ptr)
+				C.value_destroy(ptr)
 				return nil, err
 			}
 
@@ -102,7 +102,7 @@ func NewValue(val interface{}) (*Value, error) {
 			for _, key := range v.MapKeys() {
 				kv, err := NewValue(v.MapIndex(key).Interface())
 				if err != nil {
-					C._value_destroy(ptr)
+					C.value_destroy(ptr)
 					return nil, err
 				}
 
@@ -131,7 +131,7 @@ func NewValue(val interface{}) (*Value, error) {
 
 			fv, err := NewValue(v.Field(i).Interface())
 			if err != nil {
-				C._value_destroy(ptr)
+				C.value_destroy(ptr)
 				return nil, err
 			}
 
@@ -143,7 +143,7 @@ func NewValue(val interface{}) (*Value, error) {
 	case reflect.Invalid:
 		C.value_set_null(ptr)
 	default:
-		C._value_destroy(ptr)
+		C.value_destroy(ptr)
 		return nil, fmt.Errorf("unable to create value of unknown type '%T'", val)
 	}
 
@@ -158,11 +158,11 @@ func NewValueFromPtr(val unsafe.Pointer) (*Value, error) {
 
 	ptr, err := C.value_new()
 	if err != nil {
-		return nil, fmt.Errorf("unable to create new PHP value" + err.Error())
+		return nil, fmt.Errorf("unable to CREATE new PHP value" + err.Error())
 	}
 
 	if _, err := C.value_set_zval(ptr, (*C.zval)(val)); err != nil {
-		return nil, fmt.Errorf("unable to set PHP value from pointer " + err.Error())
+		return nil, fmt.Errorf("unable to SET PHP value from pointer " + err.Error())
 	}
 
 	return &Value{value: ptr}, nil
@@ -277,6 +277,6 @@ func (v *Value) Destroy() {
 		return
 	}
 
-	C._value_destroy(v.value)
+	C.value_destroy(v.value)
 	v.value = nil
 }
